* src/SDCCast.c: fix bug #2873 * support/regression/tests/literalop.c: add casts to fix   integer promotion of hexadeciaml constants in ?: operator.
git-svn-id: http://svn.code.sourceforge.net/p/sdcc/code/trunk/sdcc@10979 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ support/regression/tests/literalop.c
@@ -79,12 +79,12 @@ testOpOp (void)
 
   ASSERT ((stype         ) 0xfffffff8 * (stype         ) 0xfffffff7 == 72);
 
-  ASSERT ((signed char ) -1 * (unsigned char ) 0xfffffff7 == (sizeof(int) == 2 ? 0xff09 : 0xffffff09));
+  ASSERT ((signed char ) -1 * (unsigned char ) 0xfffffff7 == (sizeof(int) == 2 ? (int)0xff09 : (int)0xffffff09));
   ASSERT ((signed short) -1 * (unsigned short) 0xfffffff7 == (sizeof(int) == 2 ?     9u : 0xffff0009));
   ASSERT ((signed LONG ) -1 * (unsigned LONG ) 0xfffffff7 == 9u);
 
-  ASSERT ((signed char ) -2 * (unsigned char ) 0x8004 == (sizeof(int) == 2 ? 0xfff8 : 0xfffffff8));
-  ASSERT ((signed short) -2 * (unsigned short) 0x8004 == (sizeof(int) == 2 ? 0xfff8 : 0xfffefff8));
+  ASSERT ((signed char ) -2 * (unsigned char ) 0x8004 == (sizeof(int) == 2 ? (int)0xfff8 : (int)0xfffffff8));
+  ASSERT ((signed short) -2 * (unsigned short) 0x8004 == (sizeof(int) == 2 ? (int)0xfff8 : (int)0xfffefff8));
   ASSERT ((signed LONG ) -2 * (unsigned LONG ) 0x8004 == 0xfffefff8);
 
   ASSERT (-1 * 0xfff7 == (sizeof(int) == 2 ? 9 : 0xffff0009)); // 0xfff7 is stored in 'unsigned int'
@@ -134,12 +134,12 @@ testOpOp (void)
 
   ASSERT ((stype         ) 0xfffffff8 * (stype         ) 0xfffffff7 == 72);
 
-  ASSERT ((signed char ) -1 * (unsigned char ) 0xfffffff7 == (sizeof(int) == 2 ? 0xff09 : 0xffffff09));
+  ASSERT ((signed char ) -1 * (unsigned char ) 0xfffffff7 == (sizeof(int) == 2 ? (int)0xff09 : (int)0xffffff09));
   ASSERT ((signed short) -1 * (unsigned short) 0xfffffff7 == (sizeof(int) == 2 ?     9u : 0xffff0009));
   ASSERT ((signed LONG ) -1 * (unsigned LONG ) 0xfffffff7 == 9u);
 
-  ASSERT ((signed char ) -2 * (unsigned char ) 0x8004 == (sizeof(int) == 2 ? 0xfff8 : 0xfffffff8));
-  ASSERT ((signed short) -2 * (unsigned short) 0x8004 == (sizeof(int) == 2 ? 0xfff8 : 0xfffefff8));
+  ASSERT ((signed char ) -2 * (unsigned char ) 0x8004 == (sizeof(int) == 2 ? (int)0xfff8 : (int)0xfffffff8));
+  ASSERT ((signed short) -2 * (unsigned short) 0x8004 == (sizeof(int) == 2 ? (int)0xfff8 : (int)0xfffefff8));
   ASSERT ((signed LONG ) -2 * (unsigned LONG ) 0x8004 == 0xfffefff8);
 
   /* div ast: valDiv() */
@@ -147,7 +147,7 @@ testOpOp (void)
   ASSERT ((stype) -12 / (stype)  3 == (stype) -4);
   ASSERT ((stype)  12 / (stype) -3 == (stype) -4);
 
-  ASSERT ((unsigned char ) -12 / (signed char ) -3 == (sizeof(int) == 2 ? 0xffaf : 0xffffffaf));
+  ASSERT ((unsigned char ) -12 / (signed char ) -3 == (sizeof(int) == 2 ? (int)0xffaf : (int)0xffffffaf));
   ASSERT ((unsigned short) -12 / (signed short) -3 == (sizeof(int) == 2 ?      0 : 0xffffaaaf));
   ASSERT ((unsigned LONG ) -12 / (signed LONG ) -3 == 0);
   ASSERT ((utype)          -12 / (stype)         3 == (stype) 0x55555551);
@@ -174,7 +174,7 @@ testOpOp (void)
   ASSERT (s / (stype) -3 == (stype) -4);
 
   uc = -12;
-  ASSERT (uc / (signed char ) -3 == (sizeof(int) == 2 ? 0xffaf : 0xffffffaf));
+  ASSERT (uc / (signed char ) -3 == (sizeof(int) == 2 ? (int)0xffaf : (int)0xffffffaf));
   us = -12;
   ASSERT (us / (signed short) -3 == (sizeof(int) == 2 ?      0 : 0xffffaaaf));
   ul = -12;
